Move FS implementations to separate packages
